python/agents/marketing-agency/tests/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
aiortc>=1.13.0
pyttsx3==2.90
faster-whisper==1.0.3
numpy==1.26.5
soundfile==0.12.1
aiofiles==23.1.0
python-multipart==0.0.6
